👌 IMPROVE: Home ui

+++ pages/items/index.vue
@@ -28,12 +28,12 @@
         class="elevation-2"
         :no-data-text="$t('noData')"
       >
-        <template v-if="$auth.user.permissions.includes('update_item') || $auth.user.permissions.includes('delete_item')" #[`item.actions`]="{ item }">
+        <template v-if="$auth.loggedIn && ($auth.user.permissions.includes('update_item') || $auth.user.permissions.includes('delete_item'))" #[`item.actions`]="{ item }">
           <div class="d-flex gap-2">
-            <v-btn v-if="$auth.user.permissions.includes('update_item')" :to="localePath(`/items/${item.id}`)" class="primary--text" elevation="0" small>
+            <v-btn v-if="$auth.loggedIn && $auth.user.permissions.includes('update_item')" :to="localePath(`/items/${item.id}`)" class="primary--text" elevation="0" small>
               {{ $t('edit') }}
             </v-btn>
-            <v-btn v-if="$auth.user.permissions.includes('delete_item')" class="error--text" elevation="0" small @click="deleteItem(item.id)">
+            <v-btn v-if="$auth.loggedIn && $auth.user.permissions.includes('delete_item')" class="error--text" elevation="0" small @click="deleteItem(item.id)">
               {{ $t('delete') }}
             </v-btn>
           </div>
